pyproject.toml
[project]
name = "sentosa"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "torch>=2.10.0",
    "transformers>=5.1.0",
]
